CodeRunner/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docker==7.0.0
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
gevent==24.2.1
google-auth==2.29.0
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
kubernetes==29.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
